refactor: Split adaptive_execution.rs into focused modules
Split the 769-line adaptive_execution.rs module into smaller, focused files:

- adaptive/mod.rs: Main module with ExecutionModel enum and public API
- adaptive/expression.rs: Expression traversal utilities (aggregate, arithmetic, window)
- adaptive/query.rs: Query characteristic analysis (GROUP BY, joins, projections)
- adaptive/patterns.rs: Analytical pattern detection logic
- adaptive/hints.rs: Query hint extraction (placeholder for future)

Benefits:
- Improved code maintainability and readability
- Better separation of concerns
- Easier to navigate and modify specific functionality
- Reduced cognitive load when working on optimization strategies

All existing tests pass. Public API remains unchanged.

Closes #2461
